CORE-6723 fix config value

+++ components/flow/flow-service/src/main/kotlin/net/corda/flow/pipeline/impl/FlowEventProcessorImpl.kt
@@ -57,7 +57,7 @@ class FlowEventProcessorImpl(
 
         //flow result timeout must be lower than the processor timeout as the processor thread will be killed by the subscription consumer
         // thread after this period and so this timeout would never be reached and given a chance to return otherwise.
-        val flowTimeout = config.getInt(PROCESSING_MAX_FLOW_EXECUTION_DURATION) * 0.75.toLong()
+        val flowTimeout = (config.getLong(PROCESSING_MAX_FLOW_EXECUTION_DURATION) * 0.75).toLong()
         return try {
             flowEventContextConverter.convert(
                 pipeline
